Conflicts:
	CMakeLists.txt
	src/components/application_manager/CMakeLists.txt
	src/components/application_manager/include/application_manager/resumption/resumption_data_db.h
	src/components/application_manager/src/commands/hmi/get_urls.cc
	src/components/application_manager/src/commands/hmi/on_received_policy_update.cc
	src/components/application_manager/src/commands/hmi/on_system_request_notification.cc
	src/components/application_manager/src/commands/hmi/sdl_policy_update.cc
	src/components/application_manager/src/policies/policy_handler.cc
	src/components/application_manager/src/resumption/resumption_data_db.cc
	src/components/application_manager/src/resumption/resumption_data_json.cc
	src/components/application_manager/test/CMakeLists.txt
	src/components/application_manager/test/hmi_capabilities_test.cc
	src/components/application_manager/test/message_helper/CMakeLists.txt
	src/components/application_manager/test/message_helper/message_helper_test.cc
	src/components/application_manager/test/mock/include/application_manager/application_manager_impl.h
	src/components/application_manager/test/mock/include/application_manager/event_engine/event_dispatcher_impl.h
	src/components/application_manager/test/resumption/resumption_data_json_test.cc
	src/components/application_manager/test/resumption/resumption_data_test.cc
	src/components/application_manager/test/state_controller/state_controller_test.cc
	src/components/connection_handler/test/CMakeLists.txt
	src/components/formatters/CMakeLists.txt
	src/components/formatters/test/cFormatterJsonSDLRPCv1_test.cc
	src/components/formatters/test/formatter_json_rpc_test.cc
	src/components/formatters/test/meta_formatter_test.cc
	src/components/formatters/test/src/create_smartSchema.cc
	src/components/formatters/test/src/meta_formatter_test_helper.cc
	src/components/hmi_message_handler/CMakeLists.txt
	src/components/include/application_manager/application_manager.h
	src/components/include/test/application_manager/mock_application_manager.h
	src/components/media_manager/CMakeLists.txt
	src/components/policy/include/policy/sql_pt_representation.h
	src/components/policy/src/cache_manager.cc
	src/components/policy/src/policy/CMakeLists.txt
	src/components/policy/src/policy/policy_table/table_struct/CMakeLists.txt
	src/components/policy/src/policy/usage_statistics/CMakeLists.txt
	src/components/policy/src/policy_manager_impl.cc
	src/components/policy/src/policy_table/types.cc
	src/components/policy/src/sql_pt_representation.cc
	src/components/policy/src/usage_statistics/counter.cc
	src/components/policy/test/CMakeLists.txt
	src/components/policy/test/include/policy/mock_policy_listener.h
	src/components/policy/test/policy_manager_impl_test.cc
	src/components/policy/test/sql_pt_representation_test.cc
	src/components/policy/test/update_status_manager_test.cc
	src/components/protocol/CMakeLists.txt
	src/components/protocol_handler/test/CMakeLists.txt
	src/components/rpc_base/CMakeLists.txt
	src/components/telemetry_monitor/test/CMakeLists.txt
	src/components/transport_manager/CMakeLists.txt
	src/components/transport_manager/test/CMakeLists.txt
	src/components/utils/CMakeLists.txt
	src/components/utils/include/utils/winhdr.h
	src/components/utils/src/sqlite_wrapper/CMakeLists.txt
	src/components/utils/src/timer.cc
	src/components/utils/test/CMakeLists.txt
	src/components/utils/test/file_system_test.cc

Related-issues: APPLINK-22713, APPLINK-22716

+++ src/3rd_party-static/MessageBroker/src/client/mb_controller.cpp
@@ -47,7 +47,7 @@ namespace NsMessageBroker
          std::string wmes = m_receiverWriter.write(root);
          DBG_MSG(("Parsed JSON string:%s; length: %d\n", wmes.c_str(), wmes.length()));
          DBG_MSG(("Buffer is:%s\n", m_receivingBuffer.c_str()));
-         ssize_t beginpos = m_receivingBuffer.find(wmes);
+         ssize_t beginpos = static_cast<ssize_t>(m_receivingBuffer.find(wmes));
          if (-1 != beginpos)
          {
             m_receivingBuffer.erase(0, beginpos + wmes.length());
@@ -174,7 +174,7 @@ namespace NsMessageBroker
       DBG_MSG(("CMessageBrokerController::getDestinationComponentName()\n"));
       std::string ret = "";
       std::string method = root["method"].asString();
-      int pos = method.find(".");
+      int pos = static_cast<int>(method.find("."));
       if (-1 != pos)
       {
          ret = method.substr(0, pos);
@@ -188,7 +188,7 @@ namespace NsMessageBroker
       DBG_MSG(("CMessageBrokerController::getMethodName()\n"));
       std::string ret = "";
       std::string method = root["method"].asString();
-      int pos = method.find(".");
+      int pos = static_cast<int>(method.find("."));
       if (-1 != pos)
       {
          ret = method.substr(pos+1);
